+++ Common/src/main/java/net/potionstudios/biomeswevegone/world/level/block/plants/bush/DesertPlantBlock.java
@@ -3,16 +3,16 @@ package net.potionstudios.biomeswevegone.world.level.block.plants.bush;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.tags.TagKey;
-import net.minecraft.world.entity.Entity;
-import net.minecraft.world.entity.EntityType;
-import net.minecraft.world.entity.InsideBlockEffectApplier;
-import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.*;
+import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.pathfinder.PathType;
 import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 public class DesertPlantBlock extends BWGPlacementBushBlock {
 
@@ -29,5 +29,15 @@ public class DesertPlantBlock extends BWGPlacementBushBlock {
                     entity.hurtServer(serverLevel, entity.damageSources().cactus(), 1.0F);
             }
         }
+    }
+
+    /**
+     * Overrides the path type for mobs walking through the bush.getBlockPathType
+     * @see net.neoforged.neoforge.common.extensions.IBlockExtension#getBlockPathType
+     * @see net.minecraftforge.common.extensions.IForgeBlock#getBlockPathType
+     */
+    @Nullable
+    PathType getBlockPathType(BlockState state, BlockGetter level, BlockPos pos, @Nullable Mob mob) {
+        return PathType.DAMAGE_OTHER;
     }
 }